add create.jsp

+++ src/main/java/com/anadeainc/zelenko/cms/repository/PageRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface PageRepository extends CrudRepository<Page, Long> {
+public interface PageRepository extends CrudRepository<Page, Integer> {
     List<Page> findAllByOrderByPriority();
 }
